click event for burges ingredients, css tune, deprecated images deletion
