+++ src/components/modal/modal.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom'
 import modalStyles from './modal.module.css';
-import meatStub from '../../images/meat-01.png';
 import IngredientDetails from '../ingredient-details/ingredient-details';
 import OrderDetails from '../order-details/order-details';
